feat(component-tester): replace ComponentTester.waitForElement by main waitFor function
