+++ src/aurelia-testing.js
@@ -1,6 +1,7 @@
 import {CompileSpy} from './compile-spy';
 import {ViewSpy} from './view-spy';
 import {StageComponent, ComponentTester} from './component-tester';
+import {waitFor, waitForDocumentElement, waitForDocumentElements} from './wait';
 
 function configure(config) {
   config.globalResources(
@@ -14,5 +15,8 @@ export {
   ViewSpy,
   StageComponent,
   ComponentTester,
-  configure
+  configure,
+  waitFor,
+  waitForDocumentElement,
+  waitForDocumentElements
 };
